requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bson==0.5.10
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-autocomplete-light==3.11.0
django-cropperjs==0.0.5
django-js-asset==2.2.0
django-mptt==0.16.0
django-nested-admin==4.0.2
django-pwa==1.1.0
django-resized==1.0.2
django-social-share==2.3.0
django-tinymce==3.5.0
django-webpush==0.3.4
firebase-admin==6.5.0
fonttools==4.51.0
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
http_ece==1.2.0
httplib2==0.22.0
idna==3.6
msgpack==1.0.8
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.9
py-vapid==1.9.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.9.0
PyJWT==2.8.0
pyparsing==3.1.2
pyphen==0.14.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2024.1
pywebpush==1.9.4
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.24.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
weasyprint==61.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xlwt==1.3.0
yarl==1.9.4
zopfli==0.2.3
